legal-helper-hackathon/legal-helper/backend/requirements.txt
 fastapi
uvicorn[standard]
langchain
langchain-community
langchain-google-genai
chromadb
sentence-transformers
google-generativeai
python-dotenv
pypdf2
python-multipart
pydantic
pydantic-settings
requests
beautifulsoup4
langdetect
groq
datasets
